Extract download-file to scripts/helpers
Move generic download-file utility to scripts/helpers/ for reuse by
other scripts. This prepares for #106 (pnpm workspaces for scripts).

- Move scripts/pluginval/download-file.ts to scripts/helpers/
- Move scripts/pluginval/download-file.test.ts to scripts/helpers/
- Update pluginval imports to use ../helpers/download-file.js
